+++ py-malef/src-base/py_malef-surfaces.h
@@ -89,7 +89,7 @@ pyMalef_Surface___dealloc__ ( _pyMalef_surfaceStruct *self ) {
 
 
 /*
- * *** malef.Surface.__new__ ***
+ * *** pyMalef.Surface.__new__ ***
  *
  * This is the function that is called when a new Surface is created.
  */
@@ -111,7 +111,7 @@ pyMalef_Surface___new__ ( PyTypeObject *type,
 
 
 /*
- * *** malef.Surface.__init__ ***
+ * *** pyMalef.Surface.__init__ ***
  *
  * This function is called when the Surface is initialized, in Python it takes
  * the following parameters:
@@ -175,7 +175,7 @@ pyMalef_Surface___finalize__ ( PyObject *self ) {
  *####################### P U B L I C   M E T H O D S #######################*
 \*###########################################################################*/
 
-/* *** malef.Surface.debugPut *** */
+/* *** pyMalef.Surface.debugPut *** */
 PyDoc_STRVAR (pyMalef_Surface_debugPut_doc,
 "This function is currently used for development purposes, it only prints the"\
 " surface onto the screen. A better function will be added in the future.") ;
@@ -194,7 +194,7 @@ pyMalef_Surface_debugPut ( _pyMalef_surfaceStruct *self,
    pyMalef_Surface_debugPut_doc                                               \
 }
 
-/* *** malef.Surface.getForeground *** */
+/* *** pyMalef.Surface.getForeground *** */
 PyDoc_STRVAR ( pyMalef_Surface_getForeground_doc,
 "This function can be used to get the foreground colour in a certain position"\
 " of the surface" ) ;
@@ -236,7 +236,7 @@ pyMalef_Surface_getForeground ( _pyMalef_surfaceStruct *self,
    pyMalef_Surface_getForeground_doc                                          \
 }
 
-/* *** malef.Surface.getBackground *** */
+/* *** pyMalef.Surface.getBackground *** */
 PyDoc_STRVAR ( pyMalef_Surface_getBackground_doc,
 "This function returns the background colour from a certain position of the " \
 "surface." ) ;
@@ -279,7 +279,7 @@ pyMalef_Surface_getBackground ( _pyMalef_surfaceStruct *self,
 
 
 
-/* *** malef.Surface.setForeground *** */
+/* *** pyMalef.Surface.setForeground *** */
 PyDoc_STRVAR ( pyMalef_Surface_setForeground_doc,
 "This function can be used to change the foreground colour in a certain "     \
 "position of the surface." ) ;
@@ -357,7 +357,7 @@ pyMalef_Surface_setForeground ( _pyMalef_surfaceStruct *self,
 
 
 
-/* *** malef.Surface.setBackground *** */
+/* *** pyMalef.Surface.setBackground *** */
 PyDoc_STRVAR ( pyMalef_Surface_setBackground_doc,
 "This function can be used to change the background colour in a certain "     \
 "position of the surface." ) ;
@@ -430,7 +430,7 @@ pyMalef_Surface_setBackground ( _pyMalef_surfaceStruct *self,
 } 
 
 
-/* *** malef.Surface.getCursorForeground *** */
+/* *** pyMalef.Surface.getCursorForeground *** */
 PyDoc_STRVAR ( pyMalef_Surface_getCursorForeground_doc,
 "This function returns the colour of the cursor, i.e. the default writing "
 "colour of the surface." ) ;
@@ -462,7 +462,7 @@ pyMalef_Surface_getCursorForeground ( _pyMalef_surfaceStruct *self,
 }
 
 
-/* *** malef.Surface.getCursorBackground *** */
+/* *** pyMalef.Surface.getCursorBackground *** */
 PyDoc_STRVAR ( pyMalef_Surface_getCursorBackground_doc,
 "This function returns the background colour of the cursor, i.e. the default "
 "writing colour of the surface on the background." ) ;
@@ -492,7 +492,7 @@ pyMalef_Surface_getCursorBackground ( _pyMalef_surfaceStruct *self,
 }
 
 
-/* *** malef.Surface.setCursorForeground *** */
+/* *** pyMalef.Surface.setCursorForeground *** */
 PyDoc_STRVAR ( pyMalef_Surface_setCursorForeground_doc,
 "This function can be used to change the default writing foreground colour "
 "of a Surface." ) ;
@@ -600,7 +600,7 @@ pyMalef_SurfaceMethods[] = {
 static PyTypeObject
 pyMalef_Surface = {
     PyVarObject_HEAD_INIT ( NULL, 0 )
-   .tp_name      = "malef.Surface",
+   .tp_name      = "pyMalef.Surface",
    .tp_basicsize = sizeof(_pyMalef_surfaceStruct),
 // .tp_itemsize
    .tp_dealloc   = (destructor)pyMalef_Surface___dealloc__,